fix size to length
in tha admin are when using the page-images plugin, an exception is thrown with: undefined method `name' for nil:NilClass. I saw that the method `size` is used in app/views/refinery/admin/pages/tabs/_images_field.html.erb. I changed it to `length` and it works. `size` and `length` create two very different queries in AR.